// src/pages/Foods/types/ForeignFoodRecallInfo.d.ts
export interface ForeignFoodRecallInfo {
    TITL:             string; // 제품명
    DETECT_TITL:      string; // 유해물질
    CRET_DTM:         string; // 생성일자
    BDT:              string; // 본문내용
    DOWNLOAD_URL:     string; // 이미지 다운로드 URL
    NTCTXT_NO:        string; // 게시글번호
}